don't use backslashes in Python interpreter path on Windows
When passing in a runtime to pip_import() that is symlinked to a
Windows path, eg:

        python_runtime = "@python//:python",

which is symlinked to c:\python\python.exe

then the backslash-containing path is included verbatim in
the generated repositories.bzl file. When Bazel tries to read
the file, it complains about unrecognized control characters.
Converting the backslashes to forward slashes fixes the error,
and allows the build to proceed.